Allow passing options via a .yml file with the path through an env var
This way the call to email::init needs no parameters. This makes it a
bit simpler and also makes integrating with other thing easier.

The path to the options file should be set in the `EMAIL_OPTIONS_FILE`
env var, and is relative tom the current working directory.

Additionally, users can set `EMAIL_DEBUG` (to anything) to enable debug
outputs.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>