Include both SMTP and IMAP hostnames in config file
This way the same config file can be used for both sending and receiving
emails.

However, it is not necessary to actually a value for  `url-imap` if we
do not want to receive emails. Same for `url-smtp` and sending emails.
This means that we have to check that the one we want to use has
actually been set.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>