Add SubscriberManager to poll for emails and give to subscribers (#50)
This moves the email polling logic (using EmailReceiver) from the
Subscriber to a central location: the SubscriberManager. Each
Subscriber has a queue of messages. It registers its topic name and its
queue with the manager, which keeps a map. The manager starts a thread
that polls for new emails. When it gets one, it checks if the
topic/subject of the email corresponds to a subscriber in its map. If
so, it pushes it to the subscriber's queue.

The SubscriberManager is started (i.e. its thread is created) by the
Context, whenever needed.

This makes the Subscriber's get_message() method be a simple check for a
new message. However, this also adds a get_message_and_wait() util
method that waits and returns only when there is a new message.

Overall, this is the initial work to allow for multiple subscribers
easily. Some TODOs track various ideas for improvements.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>