Add ServiceClient & ServiceServer (#59)
Initial implementation; improvements to follow.

The ServiceClient has a Publisher and a Subscriber. It sends its request
using its Publisher, and gets the response using its Subscriber.

The ServiceServer has a Publisher and a special requests queue. Its
Publisher is used to send back its response as an email reply. It
registers its requests queue with the SubscriberManager (to be
refactored) in order to keep the whole EmailData struct, not just the
email body.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>