Split SubscriberManager into PollingManager and Handler classes (#63)
PollingManager does the polling using the EmailReceiver. It has a vector
for handlers to register with it. When it gets new emails, it calls
every handler with the new email.

SubscriptionHandler and ServiceHandler both register with the
PollingManager. Subscribers/service servers register with the right
handler, which pushes the new message or new email data to their queues
if there is a match.

Overall, this splits up the polling and handling logic so that the
polling logic doesn't need to contain handling logic, like giving the
new message to subscribers depending on the topic/email subject.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>