Use logger (#68)
This adds logging using spdlog. A "log" header provides a couple
functions to abstract away the actual logger implementation. The root
logger is initialized as part of the context initialization.

This replaces the "debug" flag/option with a logger in most places, but
keeps the flag and renames it to "curl verbose" to control the verbose
mode for curl separately, e.g. with the "EMAIL_CURL_VERBOSE" environment
variable.

The logging level defaults to "info" and can be set using the
"EMAIL_LOG_LEVEL" environment variable, which simply maps to the spdlog
logging levels, e.g. `EMAIL_LOG_LEVEL=debug ./app`.

Curently, it only prints logs to stdout.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>